fix: DoubleBigDecimal + 乐观锁修复
TD-004: Dish/Product/GardenService/OrderItem/Order price字段从Double改为BigDecimal
TD-005: DiningService/ProductService扣库存改用updateById，乐观锁生效

+++ server/src/main/java/com/agritainment/dto/CreateProductRequest.java
@@ -3,6 +3,7 @@ package com.agritainment.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 public class CreateProductRequest {
@@ -10,8 +11,8 @@ public class CreateProductRequest {
     private String name;
     private String type;
     @NotNull(message = "价格不能为空")
-    private Double price;
-    private Double member_price;
+    private BigDecimal price;
+    private BigDecimal member_price;
     private String image_url;
     private String description;
     private Integer daily_quota;
